refactor: use new experiment design
