+++ internal/engine/experiment/portfiltering/testkeys.go
@@ -1,0 +1,8 @@
+package portfiltering
+
+import "github.com/ooni/probe-cli/v3/internal/model"
+
+// TestKeys contains the experiment results.
+type TestKeys struct {
+	TCPConnect []*model.ArchivalTCPConnectResult `json:"tcp_connect"`
+}
